Grid, hover effect and randomized color
Grid setup using nested for loops. The first loop creates the row and appends to the container. The second adds the columns and appends to the row.

Hover effect for drawing uses an eventlistener mouse over to change the background style of the colums.

The random rgb is a function that uses the math.random and floor function to return random values for rgb and changes every column